model-ad/MG-394: reset model details variables when route is updated to ensure default tab is loaded for each model
