+++ libs/model-ad/model-details/src/lib/model-details.component.ts
@@ -68,9 +68,16 @@ export class ModelDetailsComponent implements OnInit, AfterViewInit {
   activePanel = 'omics';
   activeParent = '';
 
+  reset() {
+    this.model = undefined;
+    this.activePanel = 'omics';
+    this.activeParent = '';
+    this.isLoading = true;
+  }
+
   ngOnInit() {
     this.route.paramMap.pipe(takeUntilDestroyed(this.destroyRef)).subscribe((params: ParamMap) => {
-      this.isLoading = true;
+      this.reset();
 
       // only fetch data during client hydration
       if (this.platformService.isBrowser) {
